control-plane/api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
httpx
python-keycloak
minio
requests
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
